Cleaned up admin and moved reporting (#337)
* Cleaned up admin and moved reporting

* Grouped files
