--- services/backend/src/core/admin/reporting/index.js
@@ -1,3 +1,0 @@
-const get = require("./get");
-
-module.exports = { get };
