--- services/backend/src/core/statistics/statistics.js
@@ -1,1 +1,0 @@
-// TODO: move reporting here
