+++ services/backend/src/core/statistics/util/flaggedProfiles.js
@@ -1,12 +1,12 @@
-const Models = require("../../../../database/models");
+const Models = require("../../../database/models");
 
 const Profiles = Models.profile;
+
 async function flaggedProfiles() {
-  const flag = await Profiles.findAll({
+  return Profiles.findAll({
     where: { flagged: true },
     attributes: ["id", "firstName", "lastName"],
   });
-  return flag;
 }
 
 module.exports = flaggedProfiles;
