+++ services/backend/src/core/statistics/util/getTopFive.js
