+++ services/backend/src/core/statistics/util/index.js
